refactor: QuestionRepository가 제네릭 기반으로 수행되지 않도록 수정
- 스프링은 인터페이스가 아닌 구현체를 기준으로 빈을 등록되기 떄문에 제네릭 타입이 다르다고 다수의 빈이 생성되지 않음
- 쓰이지않는 JpaRepository 제거
