--- persistence/src/main/java/com/comssa/persistence/question/repository/MajorDescriptiveQuestionJpaRepository.java
@@ -1,10 +1,0 @@
-package com.comssa.persistence.question.repository;
-
-import com.comssa.persistence.question.domain.major.MajorDescriptiveQuestion;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MajorDescriptiveQuestionJpaRepository extends JpaRepository<MajorDescriptiveQuestion, Long> {
-
-}
