--- persistence/src/main/java/com/comssa/persistence/question/repository/jpa/MajorQuestionChoiceRepository.java
@@ -1,10 +1,0 @@
-package com.comssa.persistence.question.repository.jpa;
-
-
-import com.comssa.persistence.question.domain.major.MajorQuestionChoice;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MajorQuestionChoiceRepository extends JpaRepository<MajorQuestionChoice, Long> {
-}
