--- persistence/src/main/java/com/comssa/persistence/question/repository/jpa/QuestionCustomRepository.java
@@ -1,10 +1,0 @@
-package com.comssa.persistence.question.repository.jpa;
-
-
-import com.comssa.persistence.question.domain.common.Question;
-
-import java.util.Optional;
-
-public interface QuestionCustomRepository<T extends Question> {
-	Optional<T> findByIdFetchCommentsOrderByCreatedAtDesc(Long questionId);
-}
