+++ persistence/src/main/java/com/comssa/persistence/question/service/QuestionRepositoryService.java
@@ -1,7 +1,7 @@
 package com.comssa.persistence.question.service;
 
 import com.comssa.persistence.question.domain.common.Question;
-import com.comssa.persistence.question.repository.querydsl.QuestionRepository;
+import com.comssa.persistence.question.repository.jpa.QuestionRepository;
 import lombok.RequiredArgsConstructor;
 import org.jetbrains.annotations.Nullable;
 import org.springframework.stereotype.Service;
